+++ src/core/requests/authentication.py
@@ -215,7 +215,7 @@ def http_auth_cracker(url, realm, http_request_method):
     err_msg = "Use the '--auth-cred' option to provide a valid pair of "
     err_msg += "HTTP authentication credentials (i.e. '--auth-cred=admin:admin') "
     err_msg += "or place another dictionary in the '"
-    err_msg += os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'txt')) + "/' directory."
+    err_msg += settings.TXT_DIR + "/' directory."
     settings.print_data_to_stdout(settings.END_LINE.LF + settings.print_critical_msg(err_msg))
     return False